Refactor setup_llm_relay into the backend class
