+++ BackendBench/scripts/main.py
@@ -7,7 +7,6 @@
 import logging
 import os
 import sys
-from typing import Dict
 
 import BackendBench.backends as backends
 import BackendBench.eval as eval
@@ -23,7 +22,6 @@ from BackendBench.suite import (
     SmokeTestSuite,
     TorchBenchTestSuite,
 )
-import traceback
 
 logger = logging.getLogger(__name__)
 
@@ -198,11 +196,11 @@ def cli(
     if backend == "llm-relay":
         llm_client = LLMRelayKernelGenerator(model=llm_model)
         backend = backends.LLMBackend(model=llm_model, llm_client=llm_client)
-        backend = setup_llm_backend(backend, llm_client, suite, llm_max_attempts)
+        backend.generate_kernels(suite, llm_max_attempts)
     elif backend == "llm":
         llm_client = LLMKernelGenerator(model=llm_model)
         backend = backends.LLMBackend(model=llm_model, llm_client=llm_client)
-        backend = setup_llm_backend(backend, llm_client, suite, llm_max_attempts)
+        backend.generate_kernels(suite, llm_max_attempts)
     elif backend == "kernel_agent":
         if backends.KernelAgentBackend is None:
             raise NotImplementedError("KernelAgent backend is for internal use only")
@@ -311,115 +309,6 @@ def cli(
             p=p,
         )
         print(f"Results saved to: {log_dir}")
-
-
-def setup_llm_backend(llm_backend, llm_client, suite, max_attempts=5):
-    """Setup LLM Relay backend by generating kernels using the local plugboard server."""
-    try:
-        successful_ops = 0
-        total_ops = 0
-        for op_test in suite:
-            op = op_test.op
-            total_ops += 1
-
-            # Extract op name more carefully - e.g., torch.ops.aten.relu.default -> relu
-            op_str = str(op)
-            if "aten." in op_str:
-                # Extract the operation name before any variant (like .default)
-                op_name = op_str.split("aten.")[-1].split(".")[0]
-            else:
-                op_name = op_str.split(".")[-1]
-
-            op_signature = f"def {op_name}(*args, **kwargs) -> torch.Tensor"
-            op_description = f"PyTorch operation: {op_name}"
-
-            print(
-                f"\n[{total_ops}] Generating kernel for {op_name} (full op: {op_str}) with up to {max_attempts} attempts"
-            )
-
-            # Create feedback callback
-            def feedback_callback(kernel_code: str, attempt: int) -> tuple[bool, Dict]:
-                # TODO: Add performance testing in addition to correctness testing
-                return llm_backend.test_kernel_correctness(
-                    op, kernel_code, op_test.correctness_tests, attempt
-                )
-
-            # Generate kernel with iterative refinement
-            kernel_code, attempts_used, success = llm_client.generate_kernel_with_retry(
-                op_name,
-                op_signature,
-                op_description,
-                framework="triton",
-                max_attempts=max_attempts,
-                feedback_callback=feedback_callback,
-            )
-            llm_backend.add_kernel(op, kernel_code, op_name)
-            if success:
-                print(
-                    f"✓ Successfully generated and compiled kernel for {op_name} after {attempts_used} attempts"
-                )
-                successful_ops += 1
-
-                # Save summary of this operation
-                summary_file = os.path.join(llm_backend.kernels_dir, f"{op_name}_summary.txt")
-                with open(summary_file, "w") as f:
-                    f.write(f"Operation: {op_name}\n")
-                    f.write(f"Full op: {op_str}\n")
-                    f.write(f"Attempts used: {attempts_used}/{max_attempts}\n")
-                    f.write("Final status: Success\n")
-                    f.write(f"Model: {llm_client.model}\n")
-                    f.write(f"Server: {llm_client.readme_server_description}\n")
-                    f.write(f"Final kernel file: {op_name}_kernel_attempt_{attempts_used}.py\n")
-
-            if not success:
-                # Save summary of this operation
-                summary_file = os.path.join(llm_backend.kernels_dir, f"{op_name}_summary.txt")
-                with open(summary_file, "w") as f:
-                    f.write(f"Operation: {op_name}\n")
-                    f.write(f"Full op: {op_str}\n")
-                    f.write(f"Attempts used: {attempts_used}/{max_attempts}\n")
-                    f.write("Final status: Failed - All attempts failed correctness tests\n")
-                    f.write(f"Model: {llm_client.model}\n")
-                    f.write(f"Server: {llm_client.readme_server_description}\n")
-                    f.write(f"Last kernel file: {op_name}_kernel_attempt_{attempts_used}.py\n")
-                # Continue with other operations
-
-        failed_ops = total_ops - successful_ops
-        success_rate = f"{successful_ops / total_ops * 100:.1f}%" if total_ops > 0 else "0.0%"
-        separator_line = "=" * 60
-        # Console output format
-        output_lines = [
-            f"\n{separator_line}",
-            "LLM BACKEND SETUP SUMMARY",
-            separator_line,
-            f"Total operations attempted: {total_ops}",
-            f"Successfully created correct kernels for: {successful_ops} ops",
-            f"Failed to create correct kernels for: {failed_ops} ops",
-            f"Success rate: {success_rate}",
-            f"Model used: {llm_client.model}",
-            f"Server: {llm_client.readme_server_description}",
-            f"Generated kernels saved to: {llm_backend.kernels_dir}",
-            f"Backend: LLM \n{separator_line}\n",
-        ]
-
-        # Print summary
-        for line in output_lines:
-            print(line)
-
-        # Save overall summary
-        overall_summary_file = os.path.join(llm_backend.kernels_dir, "OVERALL_SUMMARY.txt")
-        with open(overall_summary_file, "w") as f:
-            f.write("\n".join(output_lines))
-
-        return llm_backend
-
-    except Exception as e:
-        print(f"Error setting up LLM backend: {e}")
-        print(traceback.format_exc())
-        if "Cannot connect to LLM server" in str(e):
-            print("Please start the server please see the setup instructions below:")
-            print(f"{llm_client.readme_setup_section}")
-        sys.exit(1)
 
 
 def setup_kernel_agent_backend(kernel_agent_backend, suite, num_workers=4, max_rounds=10):
